backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile --output-file=backend/requirements.txt backend/requirements.in
#
amqp==5.3.1
    # via kombu
asgiref==3.11.1
    # via
    #   django
    #   django-cors-headers
billiard==4.2.4
    # via celery
boolean-py==5.0
    # via license-expression
cachecontrol[filecache]==0.14.4
    # via
    #   cachecontrol
    #   pip-audit
celery==5.6.2
    # via -r backend/requirements.in
certifi==2026.1.4
    # via requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
cyclonedx-python-lib==11.6.0
    # via pip-audit
defusedxml==0.7.1
    # via py-serializable
django==6.0.2
    # via
    #   -r backend/requirements.in
    #   django-cors-headers
    #   djangorestframework
    #   djangorestframework-simplejwt
django-cors-headers==4.9.0
    # via -r backend/requirements.in
djangorestframework==3.16.1
    # via
    #   -r backend/requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.5.1
    # via -r backend/requirements.in
filelock==3.21.2
    # via cachecontrol
idna==3.11
    # via requests
iniconfig==2.3.0
    # via pytest
kombu==5.6.2
    # via celery
license-expression==30.4.4
    # via cyclonedx-python-lib
markdown-it-py==4.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
msgpack==1.1.2
    # via cachecontrol
numpy==2.4.2
    # via pgvector
packageurl-python==0.17.6
    # via cyclonedx-python-lib
packaging==26.0
    # via
    #   kombu
    #   pip-audit
    #   pip-requirements-parser
    #   pytest
pgvector==0.4.2
    # via -r backend/requirements.in
pip-api==0.0.34
    # via pip-audit
pip-audit==2.10.0
    # via -r backend/requirements.in
pip-requirements-parser==32.0.1
    # via pip-audit
platformdirs==4.7.1
    # via pip-audit
pluggy==1.6.0
    # via pytest
prompt-toolkit==3.0.52
    # via click-repl
psycopg[binary]==3.3.2
    # via -r backend/requirements.in
psycopg-binary==3.3.2
    # via psycopg
py-serializable==2.1.0
    # via cyclonedx-python-lib
pygments==2.19.2
    # via
    #   pytest
    #   rich
pyjwt==2.11.0
    # via djangorestframework-simplejwt
pyparsing==3.3.2
    # via pip-requirements-parser
pytest==9.0.2
    # via
    #   -r backend/requirements.in
    #   pytest-django
pytest-django==4.11.1
    # via -r backend/requirements.in
python-dateutil==2.9.0.post0
    # via celery
redis==7.1.1
    # via -r backend/requirements.in
requests==2.32.5
    # via
    #   cachecontrol
    #   pip-audit
rich==14.3.2
    # via pip-audit
ruff==0.15.1
    # via -r backend/requirements.in
six==1.17.0
    # via python-dateutil
sortedcontainers==2.4.0
    # via cyclonedx-python-lib
sqlparse==0.5.5
    # via django
tomli==2.4.0
    # via pip-audit
tomli-w==1.2.0
    # via pip-audit
tzdata==2025.3
    # via kombu
tzlocal==5.3.1
    # via celery
urllib3==2.6.3
    # via requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.6.0
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# pip
